+++ src/week_01/lokmanSolution/Task1Solution.java
@@ -1,4 +1,4 @@
-package week_01;
+package week_01.lokmanSolution;
 
 public class Task1Solution {
